code format and remove GenTFResources
