+++ internal/meta/meta_query.go
@@ -50,18 +50,23 @@ func (meta *MetaQuery) ListResource(ctx context.Context) (ImportList, error) {
 	if err != nil {
 		return nil, err
 	}
-	log.Printf("[DEBUG] Populate resource set")
-	if err := rset.PopulateResource(); err != nil {
-		return nil, fmt.Errorf("tweaking single resources in the azure resource set: %v", err)
-	}
-	log.Printf("[DEBUG] Reduce resource set")
-	if err := rset.ReduceResource(); err != nil {
-		return nil, fmt.Errorf("tweaking across resources in the azure resource set: %v", err)
-	}
+	var rl []resourceset.TFResource
+	if meta.useAzAPI() {
+		log.Printf("[DEBUG] Azure Resource set map to TF resource set")
+		rl = rset.ToTFAzAPIResources()
+	} else {
+		log.Printf("[DEBUG] Populate resource set")
+		if err := rset.PopulateResource(); err != nil {
+			return nil, fmt.Errorf("tweaking single resources in the azure resource set: %v", err)
+		}
+		log.Printf("[DEBUG] Reduce resource set")
+		if err := rset.ReduceResource(); err != nil {
+			return nil, fmt.Errorf("tweaking across resources in the azure resource set: %v", err)
+		}
 
-	log.Printf("[DEBUG] Azure Resource set map to TF resource set")
-	// rl := rset.ToTFResources(meta.parallelism, meta.azureSDKCred, meta.azureSDKClientOpt)
-	rl := meta.GenTFResources(rset)
+		log.Printf("[DEBUG] Azure Resource set map to TF resource set")
+		rl = rset.ToTFAzureRMResources(meta.parallelism, meta.azureSDKCred, meta.azureSDKClientOpt)
+	}
 
 	var l ImportList
 	for i, res := range rl {
